==> ADHD_script/S4_association_ADHDsymp_SCstrength/S1_ADHDsymp_SCstrengthDeviation_ABCD.Rmd <==
---
title: "V1_ADHDsymp_SCstrengthDeviation"
author: "Xiaoyu Xu"
date: "2025-05-05"
output: html_document
---

This script test whether deviations of SC strength associates with normalized symptoms within ADHD?


```{r setup, include=FALSE}
rm(list=ls())
library(mgcv);
library(parallel)
library(tidyverse)
library(psych)
library(chorddiag)
library(circlize)
library(RColorBrewer)
# set resolution
Yeoresolution <- 17
if (Yeoresolution == 7){
  Yeoresolution.delLM = 6
}else if (Yeoresolution == 17){
  Yeoresolution.delLM = 15
}
element_num <- Yeoresolution.delLM*(Yeoresolution.delLM+1)/2
# input directory
wd <- getwd()
homepath <- str_split_i(wd, "Normative_model", 1)
demopath <- paste0(homepath, '/Normative_model/demography')
interfileFolder <- paste0(homepath, '/Normative_model/interfileFolder_ABCD')
functionFolder <- paste0(homepath, "/Normative_model/functions")
resultFolder <- paste0(homepath, "/Normative_model/results_ABCD")
functionFolder.SCDev <- paste0(homepath, "/SC_development/Rcode_SCdevelopment/gamfunction")
FigureFolder <- paste0(homepath, '/Normative_model/Figures_ABCD/Yeo', Yeoresolution,'/CV75')
source(paste0(functionFolder.SCDev, "/gamminteraction.R"))
source(paste0(functionFolder.SCDev, "/SCrankcorr.R"))
source(paste0(functionFolder, "/plotmatrix.R"))
source(paste0(functionFolder, "/index2network.R"))
# Load data
SCdata.sum75.merge.ADHDTD <- readRDS(paste0(interfileFolder, '/SCdata_Yeo',Yeoresolution, '_CV75_sumSCinvnode.deviations_TDtest_ADHD.rds'))
SCdata.sum75.merge.ADHDTD$sex <- as.factor(SCdata.sum75.merge.ADHDTD$sex)
SCdata.sum75.merge.ADHDTD$meanFD <- as.numeric(SCdata.sum75.merge.ADHDTD$meanFD)
SCdata.sum75.merge.ADHDTD$if_TD <- factor(SCdata.sum75.merge.ADHDTD$if_TD, levels = c("TD", "ADHD"))
Behavior <- read.csv(paste0(demopath, "/demo_sublist7.csv"))

SCdata.ADHD <- SCdata.sum75.merge.ADHDTD %>% filter(if_TD=="ADHD")
SCdata.TD <- SCdata.sum75.merge.ADHDTD %>% filter(if_TD=="TD")

sigedge.all <- read.csv(paste0(resultFolder, "/SCDeviationChange_gamresults_ADHD_sig.csv"))
sigedge.all <- sigedge.all$parcel

```

## Step1 Associations between deviations and symptoms

1. SC_Deviation ~ Symptom + s(age) + sex

```{r associate}
dataname <- "SCdata.ADHD"

symptomvar <- c("cbcl_scr_dsm5_adhd_t","cbcl_scr_syn_attention_t")

for (VOI in symptomvar){
  if (!file.exists(paste0(interfileFolder, '/SCDeviation_', VOI,'_Age_CV75_Yeo', Yeoresolution,'_ADHDall_controlAgeSex.rds'))){
    print(VOI)
    resultsum <- mclapply(1:element_num, function(x){
      region<-paste0("SC.", x, "_deviationZ")

      gamresult<-gamm.smooth.predict.covariateinteraction(region=VOI, dataname, smooth_var="age", int_var=region, int_var.predict.percentile=0.1, covariates="sex+meanFD", knots=3, set_fx = T, increments=100, stats_only=T, if_residual=FALSE)
      gamresult<-as.data.frame(gamresult)
      
      return(gamresult)
    }, mc.cores = 40)
    
    resultsum.df <- do.call(rbind, lapply(resultsum, function(x) data.frame(x)))
    resultsum.df[,c(3:18)]<-lapply(resultsum.df[,c(3:18)], as.numeric)
    
    saveRDS(resultsum.df, paste0(interfileFolder, '/SCDeviation_', VOI,'_Age_CV75_Yeo', Yeoresolution,'_ADHDall_controlAgeSex.rds'))
  }
}


```


## Step2 plot the associations

```{r plot_association}

SC_symptom.list <- list()
SCcorr.df.list <- list()
j=0
for (VOI in symptomvar){
  resultsum.df <- readRDS(paste0(interfileFolder, '/SCDeviation_', VOI,'_Age_CV75_Yeo', Yeoresolution,'_ADHDall_controlAgeSex.rds'))
  SC_symptom.list[[VOI]] <- resultsum.df
  ## All 120 edges
  # symptom effects
  resultsum.df$bootstrap.P.fdr <- p.adjust(resultsum.df$bootstrap.P.disease, method="fdr")
  numsig.fdr <- sum(resultsum.df$bootstrap.P.fdr < 0.05)
  print(paste("After FDR, deviation of", numsig.fdr, "edges are significantly associated with", VOI, "within all edges."))
  
  SCcorr.df <- SCrankcorr(resultsum.df, "T.disease", Yeoresolution.delLM)
  SCcorr.df$symptom <- VOI
  j=j+1
  SCcorr.df.list[[j]] <- SCcorr.df
  
  ## Edges with significant age-related deviation decrease
  
  resultsum.df.sigedge.all <- resultsum.df %>% filter(int_var %in% sigedge.all)
  resultsum.df.sigedge.all$bootstrap.P.fdr <- p.adjust(resultsum.df.sigedge.all$bootstrap.P.disease, method="fdr")
  numsig.fdr2 <- sum(resultsum.df.sigedge.all$bootstrap.P.fdr < 0.05)
  print(paste("After FDR, deviation of", numsig.fdr2, "edges are significantly associated with", VOI, "within group-differed edges during any period."))
  
  ### plot matrix
  # all elements
  lmthr <- max(abs(resultsum.df$T.disease), na.rm = T)
  axeslabels=c("VS.P", "SM.A", "VS.C", "SM.B", "DA.A", "DA.B", "FP.C", "DM.C", "VA.A", "TP", "FP.A", "VA.B", "DM.A", "FP.B", "DM.B")
  matfig0 <- plotmatrix("resultsum.df", "T.disease", Yeoresolution.delLM, Pvar="bootstrap.P.fdr", NAcol="white", lmthr=lmthr, axeslabels=axeslabels, axeslabelsGap=T, linerange_frame=NA, PaletteSet=NA, Pvar.noFDR=NA)
  ggsave(paste0(FigureFolder, "/SCDeviation_Symptom/Matrix_SC_", VOI, "_association_all_controlAgeSex.tiff"), matfig0, height=18, width=20, units = "cm")
  
  #### Scatter plot
  if (numsig.fdr2 > 0){
    if (str_detect(VOI, "adhd")){
      ytitle <- "ADHD symptom scores"
    }else if (str_detect(VOI, "attention")){
      ytitle <- "Attention symptom scores"
    }else if (str_detect(VOI, "external")){
      ytitle <- "External symptom scores"
    }else{ytitle <- VOI}
    
    for (i in 1:numsig.fdr2){
      resultsum.df.sigedge.all.fdr <- resultsum.df.sigedge.all %>% filter(bootstrap.P.fdr < 0.05)
      region <- resultsum.df.sigedge.all.fdr$int_var[i]
      
      residual.df <- gamm.smooth.predict.covariateinteraction(region, dataname, smooth_var="age", int_var=VOI, int_var.predict.percentile=0.1, covariates="sex", knots=3, set_fx = T, increments=100, stats_only=F, if_residual=T)
      residual.df$subID <- str_split_i(residual.df$scanID, "_ses", 1)
      
      subtitle <- bquote(
        italic("T") == .(round(resultsum.df.sigedge.all.fdr$T.disease[i], 2)) * "," ~ 
          italic(P) == .(round(resultsum.df.sigedge.all.fdr$bootstrap.P.fdr[i], 3))
      )
      
      scatterplot <- ggplot(data=residual.df)+
        geom_line(aes(x=region.res, y=int_var.res, group=subID), color="gray", alpha=0.3)+
        geom_point(aes(x=region.res, y=int_var.res), color="#de77ae")+
        geom_smooth(aes(x=region.res, y=int_var.res), fill="#de77ae",color="black", method="lm")+
        labs(x="SC Deviation", y=ytitle)+
        theme_classic()+
        theme(axis.text=element_text(size=20.5, color="black"), 
              axis.title =element_text(size=20.5, color="black"),
              aspect.ratio = 0.95,
              plot.subtitle = element_text(size=20.5, hjust=0.1, vjust=-6, color="black"),
              plot.background=element_rect(fill="transparent"),
              panel.background=element_rect(fill="transparent"),
              legend.text = element_text(size=15, color="black"),
              legend.title = element_text(size=15, color="black"))
      
      ggsave(paste0(FigureFolder, "/SCDeviation_Symptom/Scatter/", region, "_", VOI, "_sigDiagnosis_controlAgeSex.tiff"), scatterplot, width = 13, height=13, units = "cm")
    }
  }
}

SCcorr.df <- do.call(rbind, SCcorr.df.list)

print(SCcorr.df)

symass.ADHD <- SC_symptom.list[["cbcl_scr_dsm5_adhd_t"]]
symass.ATT <- SC_symptom.list[["cbcl_scr_syn_attention_t"]]

symass.all <- rbind(symass.ADHD, symass.ATT) %>% filter(int_var %in% sigedge.all)
symass.all$bootstrap.P.fdr <- p.adjust(symass.all$bootstrap.P.disease, method = "fdr")

write.csv(symass.all, paste0(resultFolder, "/SCDeviation_Symptom_Age_CV75_Yeo", Yeoresolution,"_ADHDall_controlAgeSex.csv"), row.names = F)

symass.all.sig <- symass.all %>% filter(bootstrap.P.fdr < 0.05)
summary(symass.all$correstimate)
#     Min.  1st Qu.   Median     Mean  3rd Qu.     Max. 
# -0.0518566 -0.0007238  0.0250528  0.0250324  0.0423596  0.1002908  

# chord diagram
Yeo17_SAtab <- read.csv(paste0(resultFolder, '/Yeo17S400_SArank.csv'))
Yeo17_SAtab$index <- c(0:17)
limbicindex <- Yeo17_SAtab$index[str_detect(Yeo17_SAtab$NetworkLabel, "Limbic")]
Yeo17_SAtab <- Yeo17_SAtab %>% filter(! index %in% limbicindex)
Yeo17_SAtab$RGB <- c("white", "#771183", "#FF0200", "#4680B3", "#2ACCA5", "#499C3D", "#007512", "#C334F4", 
                            "#F997D8", "#778CB3", "#E89321", "#8A314B", "#0D28FD", "#000083", "#F9FF04", "#D13F52")
                            
Yeo17_SAtab$NetworkLabe.short <- c("None", "VS.C", "VS.P", "SM.A", "SM.B", "DA.A", "DA.B", "VA.A", "VA.B", "FP.C", "FP.A", "FP.B", "TP", "DM.C", "DM.A", "DM.B")      
Yeo17_SAtab <- Yeo17_SAtab[order(Yeo17_SAtab$MedianSA),]
           
for (VOI in c("cbcl_scr_dsm5_adhd_t", "cbcl_scr_syn_attention_t")){
  symass.tmp <- SC_symptom.list[[VOI]]
  
  symass.tmp <- symass.tmp %>% mutate(
    correstimate2 = case_when(
      int_var %in% sigedge.all ~ correstimate,
      .default = NA
    ))
  CorrMat <- matrix(NA, Yeoresolution.delLM, Yeoresolution.delLM)
  
  CorrMat[lower.tri(CorrMat, diag=T)] <- symass.tmp$correstimate2
  CorrMat[upper.tri(CorrMat)] <- t(CorrMat)[upper.tri(CorrMat)]
  delidx.row <- which(rowSums(is.na(CorrMat))==Yeoresolution.delLM)
  CorrMat <- CorrMat[-delidx.row, -delidx.row]
  
  rownames(CorrMat) <- colnames(CorrMat) <- axeslabels[setdiff(1:Yeoresolution.delLM, delidx.row)]
  CorrMat[upper.tri(CorrMat)] <- NA
  grid.col <- setNames(Yeo17_SAtab$RGB[Yeo17_SAtab$NetworkLabe.short %in% rownames(CorrMat)], rownames(CorrMat))
  col_fun <- colorRamp2(seq(from=min(symass.all$correstimate, na.rm=T), to=max(symass.all$correstimate, na.rm=T), length.out=7), brewer.pal(9, "Purples")[2:8])
  
  CorrMat2 <- CorrMat
  CorrMat2[!is.na(CorrMat2)] <- 1
  
  ColMat <- col_fun(CorrMat)
  
  tiff(paste0(FigureFolder, "/SCDeviation_Symptom/ChordPlot_Corr_", VOI,".tiff"), width = 15, height = 14, units = "cm", res = 300)
  circos.clear()
  circos.par(start.degree = 75)
  chordDiagram(x=CorrMat2, grid.col = grid.col, grid.border=NULL, col=ColMat, order=rownames(CorrMat),transparency = 0.5, directional = 0, self.link=1, link.lwd = 2, annotationTrack = "grid", preAllocateTracks = list(track.height = 0.12), big.gap=T)
  circos.track(track.index = 1, panel.fun = function(x, y) {
    sector.name <- CELL_META$sector.index
    if (sector.name %in% c("SM.B", "VS.P", "DM.A")) {
      
      circos.text(CELL_META$xcenter, CELL_META$ylim[1] + mm_y(8), sector.name,
                  facing = "clockwise", niceFacing = TRUE, cex = 1.5)
    } else {
      circos.text(CELL_META$xcenter, CELL_META$ylim[1] + mm_y(2), sector.name,
                  facing = "outside", niceFacing = TRUE, cex = 1.5)
    }
  }, bg.border = NA)
  
  dev.off()
  
}


```
