==> ADHD_script/S4_association_ADHDsymp_SCstrength/S3_delta_deviation_delta_symptom_ABCD.qmd <==
---
title: "S1_delta_deviation_delta_symptom"
author:
  - name: Xiaoyu Xu
    orcid: 0000-0002-7349-3366
    email: xuxiaoyu@cibr.ac.cn
    url: https://scholar.google.com/citations?user=G0jeI1AAAAAJl
    affiliation: 
      - name: Chinese Institute for Brain Research, Beijing (CIBR)
        group: CUIZAIXU Lab
        url: https://cuilab.cibr.ac.cn/
date: "2025-06-09"
format: 
  html:
    code-fold: true
    code-tools: true
    theme: Materia
    toc: true
    toc_depth: 3
    toc-expand: 1
    toc_float: true
    toc-location: left
    number-sections: true
    number-depth: 3
    embed-resources: true
---

This script is to test whether the SC deviation change associated with symtom change.

```{r}
#| label: load-packages-data
#| include: false
rm(list=ls())
library(ggplot2)
library(RColorBrewer)
library(tidyverse)
library(mgcv)
library(ppcor)
library(openxlsx)
library(parallel)
library(ecostats)
library(psych)
library(reshape)
library(tableone)
library(chorddiag)
library(circlize)
# set resolution
Yeoresolution <- 17
if (Yeoresolution == 7){
  Yeoresolution.delLM = 6
}else if (Yeoresolution == 17){
  Yeoresolution.delLM = 15
}
element_num <- Yeoresolution.delLM*(Yeoresolution.delLM+1)/2
# input directory
wd <- getwd()
homepath <- str_split_i(wd, "Normative_model", 1)
demopath <- paste0(homepath, '/Normative_model/demography')
interfileFolder <- paste0(homepath, '/Normative_model/interfileFolder_ABCD')
functionFolder <- paste0(homepath, "/Normative_model/functions")
resultFolder <- paste0(homepath, "/Normative_model/results_ABCD")
functionFolder.SCDev <- paste0(homepath, "/SC_development/Rcode_SCdevelopment/gamfunction")
FigureFolder <- paste0(homepath, '/Normative_model/Figures_ABCD/Yeo', Yeoresolution,'/CV75')
source(paste0(functionFolder.SCDev, "/SCrankcorr.R"))
source(paste0(functionFolder.SCDev, "/gammsmooth.R"))
source(paste0(functionFolder.SCDev, "/plotdata_generate.R"))
source(paste0(functionFolder, "/lmsymp.R"))
source(paste0(functionFolder, "/lme4symp.R"))
source(paste0(functionFolder, "/plotmatrix.R"))
source(paste0(functionFolder, "/index2network.R"))
# Load data
SCdata.sum75.merge.ADHDTD <- readRDS(paste0(interfileFolder, '/SCdata_Yeo',Yeoresolution, '_CV75_sumSCinvnode.deviations_TDtest_ADHD.rds'))
SCdata.sum75.merge.ADHDTD$sex <- as.factor(SCdata.sum75.merge.ADHDTD$sex)
SCdata.sum75.merge.ADHDTD$meanFD <- as.numeric(SCdata.sum75.merge.ADHDTD$meanFD)
SCdata.sum75.merge.ADHDTD$if_TD <- factor(SCdata.sum75.merge.ADHDTD$if_TD, levels = c("TD", "ADHD"))
Behavior <- read.csv(paste0(demopath, "/demo_sublist7.csv"))

SCdata.ADHD <- SCdata.sum75.merge.ADHDTD %>% filter(if_TD=="ADHD")
SCdata.TD <- SCdata.sum75.merge.ADHDTD %>% filter(if_TD=="TD")

SCdata.ADHDTD.l <- SCdata.sum75.merge.ADHDTD %>% group_by(subID) %>% filter(n()>1)
SCdata.ADHDTD.l$eventname <- factor(SCdata.ADHDTD.l$eventname, levels = c("baseline_year_1_arm_1", "2_year_follow_up_y_arm_1", "4_year_follow_up_y_arm_1"))
table(SCdata.ADHDTD.l$if_TD, SCdata.ADHDTD.l$eventname)
SCdata.ADHD.l <- SCdata.ADHDTD.l %>% filter(if_TD=="ADHD")

sigedge.all <- read.csv(paste0(resultFolder, "/SCDeviationChange_gamresults_ADHD_sig.csv"))
sigedge.all <- sigedge.all$parcel
Association_symptom <- read.csv(paste0(resultFolder, "/SCDeviation_Symptom_Age_CV75_Yeo17_ADHDall_controlAgeSex.csv"))

```

## Step1 Compute the deviation delta of ADHD participants.

Pair definition: (Time2 - Time1) / duration 
(1. 2Year - Base; 2. 4Year - 2Year; 3. 4Year - Base ) / duration

```{r}
#| label: Step1

symptomvar <- c("GENERAL.norm", "EXT.norm", "ADHD.norm", "INT.norm", "ATT.norm", "HI.norm", "cbcl_scr_syn_internal_t", "cbcl_scr_syn_external_t", "cbcl_scr_dsm5_adhd_t", "cbcl_scr_syn_attention_t", "cbcl_scr_syn_totprob_t")
SCdata.ADHDTD.l$visit <- factor(SCdata.ADHDTD.l$eventname, levels = c("baseline_year_1_arm_1", "2_year_follow_up_y_arm_1", "4_year_follow_up_y_arm_1"), labels=c(1, 2, 3))

# compute difference
SCdata.ADHDTD.l.delta <- SCdata.ADHDTD.l %>% dplyr::select(subID, visit, age, sex, meanFD, if_TD, all_of(symptomvar), matches("^SC.*deviationZ$")) %>%
  group_by(subID) %>% dplyr::arrange(visit, .by_group = T) %>%
  mutate(across(c(matches("^SC.*deviationZ$"), symptomvar), 
                list(diff_2_b = ~ifelse(visit == 2, (. - .[visit == 1]), NA),
                  diff_4_2 = ~ifelse(visit == 3, (. - .[visit == 2]), NA),
                  diff_4_b = ~ifelse(visit == 3, (. - .[visit == 1]), NA),
                  base = ~ifelse(visit==2 | visit==3, .[visit == 1], NA),
                  base2 = ~ifelse(visit == 3, .[visit == 2], NA)),
                .names = "{.col}_{.fn}"),
         
         across("age", 
                list(diff_2_b = ~ifelse(visit == 2, (. - .[visit == 1]), NA),
                  diff_4_2 = ~ifelse(visit == 3, (. - .[visit == 2]), NA),
                  diff_4_b = ~ifelse(visit == 3, (. - .[visit == 1]), NA)), 
                .names = "{.col}_{.fn}"),
         sex = unique(sex),
         across(c("age", "meanFD"), 
                list(mean_2_b = ~ifelse(visit == 2, (. + .[visit == 1])/2, NA),
                  mean_4_2 = ~ifelse(visit == 3, (. + .[visit == 2])/2, NA),
                  mean_4_b = ~ifelse(visit == 3, (. + .[visit == 1])/2, NA)), 
                .names = "{.col}_{.fn}")
         ) %>%
  dplyr::select(-c(matches("^SC.*deviationZ$"), all_of(symptomvar))) %>% 
  filter(visit %in% c(2,3)) %>% 
  fill(starts_with("SC_"), .direction = "downup") %>% 
  dplyr::select(-visit)

# compute difference scaled by age gap
SCdata.ADHDTD.l.delta <- SCdata.ADHDTD.l.delta %>% 
  mutate(across(matches("*_diff_2_b$"), list(delta_2b = ~ifelse(!is.na(.), . / age_diff_2_b, NA)), .names = "{.col}_{.fn}"),
         across(matches("*_diff_4_2$"), list(delta_42 = ~ifelse(!is.na(.), . / age_diff_4_2, NA)), .names = "{.col}_{.fn}"),
         across(matches("*_diff_4_b$"), list(delta_4b = ~ifelse(!is.na(.), . / age_diff_4_b, NA)), .names = "{.col}_{.fn}")
         ) %>% dplyr::select(contains("delta"), contains("age"),contains("base"), subID, contains("mean"), if_TD, sex)

index <- which(str_detect(names(SCdata.ADHDTD.l.delta), "SC.") & str_detect(names(SCdata.ADHDTD.l.delta), "delta"))
names(SCdata.ADHDTD.l.delta)[index] <- paste0(str_split_i(names(SCdata.ADHDTD.l.delta)[index], "_", 1), "_", str_split_i(names(SCdata.ADHDTD.l.delta)[index], "_", 6), "_", str_split_i(names(SCdata.ADHDTD.l.delta)[index], "_", 7))
index2 <- which(str_detect(names(SCdata.ADHDTD.l.delta), "diff") & str_detect(names(SCdata.ADHDTD.l.delta), "delta"))
names(SCdata.ADHDTD.l.delta)[index2] <- gsub("_diff_2_b", "", names(SCdata.ADHDTD.l.delta)[index2])
names(SCdata.ADHDTD.l.delta)[index2] <- gsub("_diff_4_b", "", names(SCdata.ADHDTD.l.delta)[index2])
names(SCdata.ADHDTD.l.delta)[index2] <- gsub("_diff_4_2", "", names(SCdata.ADHDTD.l.delta)[index2])

SCdata.ADHDTD.l.delta.l <- SCdata.ADHDTD.l.delta %>% 
  pivot_longer(
    cols = matches(".+_(delta_2b|delta_42|delta_4b)$"),
    names_to = c("Var", "delta_type"), 
    names_pattern = "(.+)_(delta_.*)",
    values_to = "delta_value"
  ) %>% dplyr::select(-matches("^age.*delta"))

SCdata.ADHDTD.l.delta.l2 <- SCdata.ADHDTD.l.delta.l %>% 
  pivot_wider(names_from = Var,
              values_from = delta_value,
              names_glue = "{Var}_delta")
SCdata.ADHDTD.l.delta.l2 <- SCdata.ADHDTD.l.delta.l2 %>% drop_na(ends_with("_delta"))
SCdata.ADHDTD.l.delta.l2 <- SCdata.ADHDTD.l.delta.l2 %>%
  mutate(
    across(
      matches("^SC.*_deviationZ_base$"),  # 匹配所有 SC.*_deviationZ_base 变量
      ~ if_else(
        delta_type == "delta_42",
        # 把 base2 对应列的值取出来
        get(str_replace(cur_column(), "_base$", "_base2")),
        .x
      )
    )
  )
SCdata.ADHDTD.l.delta.l2 <- SCdata.ADHDTD.l.delta.l2 %>% dplyr::select(-ends_with("_base2"))

write.csv(SCdata.ADHDTD.l.delta.l2, paste0(interfileFolder, "/Delta_SCDeviation_symptomTscore_ADHD_TD_Yeo", element_num, "_CV75.csv"), row.names = F)
# organize covariates
SCdata.ADHDTD.l.delta.l2 <- SCdata.ADHDTD.l.delta.l2 %>% mutate(
  meanFD.M = case_when(
    delta_type == "delta_2b" ~ meanFD_mean_2_b,
    delta_type == "delta_42" ~ meanFD_mean_4_2,
    delta_type == "delta_4b" ~ meanFD_mean_4_b,
    .default = NA),
  age.M = case_when(
    delta_type == "delta_2b" ~ age_mean_2_b,
    delta_type == "delta_42" ~ age_mean_4_2,
    delta_type == "delta_4b" ~ age_mean_4_b,
    .default = NA)
) %>% dplyr::select(-c(starts_with("age_"), starts_with("meanFD_")))


```

## Step2 Correlation analyses for SC.delta and symptom.delta.

delta syptom ~ delta deviation + mean age + sex + mean head motion
Referring to Chiang et al., Asian J Psychiatr, 2023.

```{r}
#| label: Step2

SCdata.ADHDTD.l.delta.l2 <- as.data.frame(SCdata.ADHDTD.l.delta.l2)

SCdata.ADHD.l.delta <- SCdata.ADHDTD.l.delta.l2 %>% filter(if_TD=="ADHD")


## Attention symptom ~ delta_SC.109
res.att.109 <- lm.fit.symp(region="SC.109_delta", dataname="SCdata.ADHD.l.delta", symp_var="cbcl_scr_syn_attention_t_delta", covariates="sex+age.M+meanFD.M", if_residual=F, IDvar="subID")
# T=2.26756, anova.P = 0.02335605, r=0.154595, corrp=0.0233752, corrmethod="pearson"
print(res.att.109)

dfresidual <- lm.fit.symp(region="SC.109_delta", dataname="SCdata.ADHD.l.delta", symp_var="cbcl_scr_syn_attention_t_delta", covariates="sex+age.M+meanFD.M", if_residual=T, IDvar="subID")

# plot
scatterplot <- ggplot(data=dfresidual)+
  geom_point(aes(x=SCDelta.res, y=SymptomDelta.res), color="#810F7C", alpha=0.3)+
  geom_smooth(aes(x=SCDelta.res, y=SymptomDelta.res), fill="#810F7C",color="black", method="lm")+
  labs(x="\u0394SC deviation of FP.A-FP.B", y="\u0394Attention symptom")+
  theme_classic()+
  theme(axis.text=element_text(size=17.5, color="black"),
        axis.title =element_text(size=17.5, color="black"),
        aspect.ratio = 0.9,
        plot.subtitle = element_text(size=20.5, hjust=0.1, vjust=-6, color="black"),
        plot.background=element_rect(fill="transparent"),
        panel.background=element_rect(fill="transparent"),
        legend.text = element_text(size=15, color="black"),
        legend.title = element_text(size=15, color="black"))
scatterplot
ggsave(paste0(FigureFolder, "/DeltaSCDeviation_DeltaSymptom/Scatter/DeltaSC109_DeltaAttention_Tscore_sigDiagnosis_controlAgeSex.tiff"), scatterplot, width = 13, height=13, units = "cm")

## ADHD symptom ~ delta_SC.119
res.att.119 <- lm.fit.symp(region="SC.119_delta", dataname="SCdata.ADHD.l.delta", symp_var="cbcl_scr_dsm5_adhd_t_delta", covariates="sex+age.M+meanFD.M", if_residual=F, IDvar="subID")
# T=2.1084975122627, anova.P = 0.0349879777, r=0.14398404, corrp=0.0348652879, corrmethod="pearson"

dfresidual <- lm.fit.symp(region="SC.119_delta", dataname="SCdata.ADHD.l.delta", symp_var="cbcl_scr_dsm5_adhd_t_delta", covariates="sex+age.M+meanFD.M", if_residual=T, IDvar="subID")

# plot
scatterplot <- ggplot(data=dfresidual)+
  geom_point(aes(x=SCDelta.res, y=SymptomDelta.res), color="#810F7C", alpha=0.3)+
  geom_smooth(aes(x=SCDelta.res, y=SymptomDelta.res), fill="#810F7C",color="black", method="lm")+
  labs(x="\u0394SC deviation of FP.B-DM.B", y="\u0394ADHD symptom")+
  theme_classic()+
  theme(axis.text=element_text(size=17.5, color="black"),
        axis.title =element_text(size=17.5, color="black"),
        aspect.ratio = 0.9,
        plot.subtitle = element_text(size=20.5, hjust=0.1, vjust=-6, color="black"),
        plot.background=element_rect(fill="transparent"),
        panel.background=element_rect(fill="transparent"),
        legend.text = element_text(size=15, color="black"),
        legend.title = element_text(size=15, color="black"))
scatterplot
ggsave(paste0(FigureFolder, "/DeltaSCDeviation_DeltaSymptom/Scatter/DeltaSC119_DeltaADHD_Tscore_sigDiagnosis_controlAgeSex.tiff"), scatterplot, width = 13, height=13, units = "cm")


p.adjust(c(0.0233752, 0.0348652879), method="fdr")
# 0.03486529 0.03486529

```
